chore: no more using lsp-config plugin and change lsp-config considering 0.11 neovim apis
