+++ init.lua
@@ -1,2 +1,3 @@
 require("core")
 require("lazy-config")
+require("lsp-config")
